+++ .env.example
@@ -4,5 +4,5 @@ DATABASE_USERNAME=postgres
 DATABASE_PASSWORD=postgres
 INTERNAL_API_KEY=internal-api-key
 API_BACKEND_BASE_URL=http://localhost:9999
-BIGQUERY_PROJECT_ID=bigquery-project-id
-BIGQUERY_CREDENTIALS=bigquery-credentials
+GCP_PROJECT_ID=google-project-id
+GCP_CREDENTIALS=google-credentials
