+++ infrastructure/github-archives-adapter/pom.xml
@@ -10,7 +10,7 @@
         <relativePath>../../pom.xml</relativePath>
     </parent>
 
-    <artifactId>big-query-adapter</artifactId>
+    <artifactId>github-archives-adapter</artifactId>
 
     <properties>
         <maven.compiler.source>17</maven.compiler.source>
